+++ packages/theme/src/public-api.ts
@@ -3,3 +3,4 @@ export * from './lib/light';
 export * from './lib/one-dark';
 export * from './lib/material-ocean';
 export * from './lib/dracula';
+export * from './lib/synthwave84';